Reduce shield penalties with Str instead of size (Lici)
To make strength a bit more meaningful for some characters that
might not currently care about it and to make tower shields more
desirable in the late game, replace the size component for the
shield penalty formula (for spell success & attack delay) with a
strength component.

With this formula, a 15 Str character will have the same penalty as
a medium-size character did before. A 21 Str character will have the
same penalty as a Formicid, Ogre or Troll did, a 9 Str character will
match an old Kobold, and a 3 Str character will match an old Spriggan.

This is a very small nerf for D:1 large species fighters and a buff
to them thereafter. Mid-sized stabbers, eg vampires, are hurt a bit
more; compensation TBD.

I vaguely suspect this commit is a bit too generous to shield users
and some nerfs are needed in some way, but I'm tossing it up for now
to think about it.