Bring back Sacks of Spiders
As an alternate to boxes of beasts. When these were removed, it was
correctly noted that they were a bit too similar to beast boxes and
phantom mirrors, and the latter were more interesting due to a lack
of spiderous variety. Well, spider reform has addressed the latter,
and item sets address the former. So let's do it again!

Reverts a6edddd4f1eafeb1ccc421d0d29b50fba57057a8.